refactor(agent): inline repeat-reply guard helpers
